+++ integration-tests/artifacts/src/pages/dynamic-dynamic.js
@@ -1,0 +1,17 @@
+import React, { Component } from "react"
+
+export default class Dynamic extends Component {
+  constructor(props) {
+    super(props)
+    this.state = { module: null }
+  }
+  componentDidMount() {
+    import(`../pages/dynamic`).then(module =>
+      this.setState({ module: module.default })
+    )
+  }
+  render() {
+    const { module: Component } = this.state
+    return <div>{Component && <Component />}</div>
+  }
+}