+++ packages/gatsby-cli/src/create-cli.ts
@@ -171,11 +171,27 @@ function buildLocalCommands(cli: yargs.Argv, isLocalSite: boolean): void {
         .option(`open-tracing-config-file`, {
           type: `string`,
           describe: `Tracer configuration file (OpenTracing compatible). See https://gatsby.dev/tracing`,
+        })
+        .option(`inspect`, {
+          type: `number`,
+          describe: `Opens a port for debugging. See https://www.gatsbyjs.org/docs/debugging-the-build-process/`,
+        })
+        .option(`inspect-brk`, {
+          type: `number`,
+          describe: `Opens a port for debugging. Will block until debugger is attached. See https://www.gatsbyjs.org/docs/debugging-the-build-process/`,
         }),
     handler: handlerP(
       getCommandHandler(`develop`, (args: yargs.Arguments, cmd: Function) => {
         process.env.NODE_ENV = process.env.NODE_ENV || `development`
         startGraphQLServer(siteInfo.directory, true)
+
+        if (args.hasOwnProperty(`inspect`)) {
+          args.inspect = args.inspect || 9229
+        }
+        if (args.hasOwnProperty(`inspect-brk`)) {
+          args.inspect = args.inspect || 9229
+        }
+
         cmd(args)
         // Return an empty promise to prevent handlerP from exiting early.
         // The development server shouldn't ever exit until the user directly
@@ -344,9 +360,29 @@ function buildLocalCommands(cli: yargs.Argv, isLocalSite: boolean): void {
   cli.command({
     command: `recipes [recipe]`,
     describe: `[EXPERIMENTAL] Run a recipe`,
+    builder: _ =>
+      _.option(`D`, {
+        alias: `develop`,
+        type: `boolean`,
+        default: false,
+        describe: `Start recipe in develop mode to live-develop your recipe (defaults to false)`,
+      }).option(`I`, {
+        alias: `install`,
+        type: `boolean`,
+        default: false,
+        describe: `Install recipe (defaults to plan mode)`,
+      }),
     handler: handlerP(
-      async ({ recipe }: yargs.Arguments<{ recipe: string | undefined }>) => {
-        await recipesHandler(siteInfo.directory, recipe)
+      async ({
+        recipe,
+        develop,
+        install,
+      }: yargs.Arguments<{
+        recipe: string | undefined
+        develop: boolean
+        install: boolean
+      }>) => {
+        await recipesHandler(siteInfo.directory, recipe, develop, install)
       }
     ),
   })