+++ packages/gatsby/src/commands/build-javascript.ts
@@ -1,8 +1,11 @@
 import { Span } from "opentracing"
 import webpack from "webpack"
+import { isEqual } from "lodash"
 import flatMap from "lodash/flatMap"
 
 import webpackConfig from "../utils/webpack.config"
+import { store } from "../redux"
+import mapTemplatesToStaticQueryHashes from "../utils/map-templates-to-static-query-hashes"
 
 import { IProgram } from "./types"
 
@@ -23,7 +26,44 @@ export const buildProductionBundle = async (
   )
 
   return new Promise((resolve, reject) => {
-    webpack(compilerConfig).run((err, stats) => {
+    const compiler = webpack(compilerConfig)
+
+    compiler.hooks.compilation.tap(`webpack-dep-tree-plugin`, compilation => {
+      compilation.hooks.seal.tap(`webpack-dep-tree-plugin`, () => {
+        const state = store.getState()
+        const mapOfTemplatesToStaticQueryHashes = mapTemplatesToStaticQueryHashes(
+          state,
+          compilation
+        )
+
+        mapOfTemplatesToStaticQueryHashes.forEach(
+          (staticQueryHashes, componentPath) => {
+            if (
+              !isEqual(
+                state.staticQueriesByTemplate.get(componentPath),
+                staticQueryHashes
+              )
+            ) {
+              store.dispatch({
+                type: `ADD_PENDING_TEMPLATE_DATA_WRITE`,
+                payload: {
+                  componentPath,
+                },
+              })
+              store.dispatch({
+                type: `SET_STATIC_QUERIES_BY_TEMPLATE`,
+                payload: {
+                  componentPath,
+                  staticQueryHashes,
+                },
+              })
+            }
+          }
+        )
+      })
+    })
+
+    compiler.run((err, stats) => {
       if (err) {
         return reject(err)
       }