+++ packages/gatsby/src/commands/develop.ts
@@ -2,17 +2,26 @@
 import path from "path"
 import http from "http"
 import tmp from "tmp"
-import { spawn } from "child_process"
+import { ChildProcess } from "child_process"
+import execa from "execa"
 import chokidar from "chokidar"
 import getRandomPort from "detect-port"
+import { detectPortInUseAndPrompt } from "../utils/detect-port-in-use-and-prompt"
 import socket from "socket.io"
 import fs from "fs-extra"
-import { isCI, slash } from "gatsby-core-utils"
-import { createServiceLock } from "gatsby-core-utils/dist/service-lock"
+import onExit from "signal-exit"
+import {
+  isCI,
+  slash,
+  createServiceLock,
+  getService,
+  updateSiteMetadata,
+  UnlockFn,
+} from "gatsby-core-utils"
 import reporter from "gatsby-cli/lib/reporter"
-import getSslCert from "../utils/get-ssl-cert"
+import { getSslCert } from "../utils/get-ssl-cert"
 import { startDevelopProxy } from "../utils/develop-proxy"
-import { IProgram } from "./types"
+import { IProgram, IDebugInfo } from "./types"
 
 // Adapted from https://stackoverflow.com/a/16060619
 const requireUncached = (file: string): any => {
@@ -57,25 +66,63 @@ const doesConfigChangeRequireRestart = (
   return true
 }
 
+// Return a user-supplied port otherwise the default Node.js debugging port
+const getDebugPort = (port?: number): number => port ?? 9229
+
+export const getDebugInfo = (program: IProgram): IDebugInfo | null => {
+  if (program.hasOwnProperty(`inspect`)) {
+    return {
+      port: getDebugPort(program.inspect),
+      break: false,
+    }
+  } else if (program.hasOwnProperty(`inspectBrk`)) {
+    return {
+      port: getDebugPort(program.inspectBrk),
+      break: true,
+    }
+  } else {
+    return null
+  }
+}
+
 class ControllableScript {
-  private process
+  private process?: ChildProcess
   private script
+  private debugInfo: IDebugInfo | null
   public isRunning
-  constructor(script) {
+  constructor(script, debugInfo: IDebugInfo | null) {
     this.script = script
+    this.debugInfo = debugInfo
   }
   start(): void {
+    const args = []
     const tmpFileName = tmp.tmpNameSync({
       tmpdir: path.join(process.cwd(), `.cache`),
     })
     fs.outputFileSync(tmpFileName, this.script)
     this.isRunning = true
-    this.process = spawn(`node`, [tmpFileName], {
+    // Passing --inspect isn't necessary for the child process to launch a port but it allows some editors to automatically attach
+    if (this.debugInfo) {
+      if (this.debugInfo.break) {
+        args.push(`--inspect-brk=${this.debugInfo.port}`)
+      } else {
+        args.push(`--inspect=${this.debugInfo.port}`)
+      }
+    }
+
+    this.process = execa.node(tmpFileName, args, {
       env: process.env,
       stdio: [`inherit`, `inherit`, `inherit`, `ipc`],
     })
   }
-  async stop(signal: string | null = null, code?: number): Promise<void> {
+  async stop(
+    signal: NodeJS.Signals | null = null,
+    code?: number
+  ): Promise<void> {
+    if (!this.process) {
+      throw new Error(`Trying to stop the process before starting it`)
+    }
+
     this.isRunning = false
     if (signal) {
       this.process.kill(signal)
@@ -90,14 +137,40 @@ class ControllableScript {
     }
 
     return new Promise(resolve => {
+      if (!this.process) {
+        throw new Error(`Trying to stop the process before starting it`)
+      }
+
       this.process.on(`exit`, () => {
-        this.process.removeAllListeners()
+        if (this.process) {
+          this.process.removeAllListeners()
+        }
+        this.process = undefined
         resolve()
       })
     })
   }
-  on(type, callback): void {
-    this.process.on(type, callback)
+  onMessage(callback: (msg: any) => void): void {
+    if (!this.process) {
+      throw new Error(`Trying to attach message handler before process started`)
+    }
+    this.process.on(`message`, callback)
+  }
+  onExit(
+    callback: (code: number | null, signal: NodeJS.Signals | null) => void
+  ): void {
+    if (!this.process) {
+      throw new Error(`Trying to attach exit handler before process started`)
+    }
+    this.process.on(`exit`, callback)
+  }
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  send(msg: any): void {
+    if (!this.process) {
+      throw new Error(`Trying to send a message before process started`)
+    }
+
+    this.process.send(msg)
   }
 }
 
@@ -107,12 +180,34 @@ let isRestarting
 const REGEX_IP = /^(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.){3}(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])$/
 
 module.exports = async (program: IProgram): Promise<void> => {
+  // In some cases, port can actually be a string. But our codebase is expecting it to be a number.
+  // So we want to early just force it to a number to ensure we always act on a correct type.
+  program.port = parseInt(program.port + ``, 10)
   const developProcessPath = slash(require.resolve(`./develop-process`))
+
+  try {
+    program.port = await detectPortInUseAndPrompt(program.port)
+  } catch (e) {
+    if (e.message === `USER_REJECTED`) {
+      process.exit(0)
+    }
+
+    throw e
+  }
+
   // Run the actual develop server on a random port, and the proxy on the program port
   // which users will access
   const proxyPort = program.port
+  const debugInfo = getDebugInfo(program)
+
+  // INTERNAL_STATUS_PORT allows for setting the websocket port used for monitoring
+  // when the browser should prompt the user to restart the develop process.
+  // This port is randomized by default and in most cases should never be required to configure.
+  // It is exposed for environments where port access needs to be explicit, such as with Docker.
+  // As the port is meant for internal usage only, any attempt to interface with features
+  // it exposes via third-party software is not supported.
   const [statusServerPort, developPort] = await Promise.all([
-    getRandomPort(),
+    getRandomPort(process.env.INTERNAL_STATUS_PORT),
     getRandomPort(),
   ])
 
@@ -141,13 +236,17 @@ module.exports = async (program: IProgram): Promise<void> => {
       )
     }
 
-    program.ssl = await getSslCert({
+    const ssl = await getSslCert({
       name: sslHost,
       caFile: program[`ca-file`],
       certFile: program[`cert-file`],
       keyFile: program[`key-file`],
       directory: program.directory,
     })
+
+    if (ssl) {
+      program.ssl = ssl
+    }
   }
 
   // NOTE(@mxstbr): We need to start the develop proxy before the develop process to ensure
@@ -158,7 +257,8 @@ module.exports = async (program: IProgram): Promise<void> => {
     program,
   })
 
-  const developProcess = new ControllableScript(`
+  const developProcess = new ControllableScript(
+    `
     const cmd = require(${JSON.stringify(developProcessPath)});
     const args = ${JSON.stringify({
       ...program,
@@ -166,22 +266,48 @@ module.exports = async (program: IProgram): Promise<void> => {
       proxyPort,
       // Don't pass SSL options down to the develop process, it should always use HTTP
       ssl: null,
+      debugInfo,
     })};
     cmd(args);
-  `)
+  `,
+    debugInfo
+  )
 
-  let unlock
+  let unlocks: Array<UnlockFn> = []
   if (!isCI()) {
-    unlock = await createServiceLock(program.directory, `developstatusserver`, {
-      port: statusServerPort,
+    const statusUnlock = await createServiceLock(
+      program.directory,
+      `developstatusserver`,
+      {
+        port: statusServerPort,
+      }
+    )
+    const developUnlock = await createServiceLock(
+      program.directory,
+      `developproxy`,
+      {
+        port: proxyPort,
+      }
+    )
+    await updateSiteMetadata({
+      name: program.sitePackageJson.name,
+      sitePath: program.directory,
+      pid: process.pid,
+      lastRun: Date.now(),
     })
 
-    if (!unlock) {
+    if (!statusUnlock || !developUnlock) {
+      const data = await getService(program.directory, `developproxy`)
+      const port = data?.port || 8000
       console.error(
-        `Looks like develop for this site is already running. Try visiting http://localhost:8000/ maybe?`
+        `Looks like develop for this site is already running, can you visit ${
+          program.ssl ? `https://` : `http://`
+        }://localhost:${port} ? If it is not, try again in five seconds!`
       )
       process.exit(1)
     }
+
+    unlocks = unlocks.concat([statusUnlock, developUnlock])
   }
 
   const statusServer = http.createServer().listen(statusServerPort)
@@ -211,30 +337,49 @@ module.exports = async (program: IProgram): Promise<void> => {
       io.emit(`develop:is-starting`)
       await developProcess.stop()
       developProcess.start()
-      developProcess.on(`message`, handleChildProcessIPC)
+      developProcess.onMessage(handleChildProcessIPC)
       isRestarting = false
     })
   })
 
   developProcess.start()
-  developProcess.on(`message`, handleChildProcessIPC)
+  developProcess.onMessage(handleChildProcessIPC)
 
   // Plugins can call `process.exit` which would be sent to `develop-process` (child process)
   // This needs to be propagated back to the parent process
-  developProcess.on(`exit`, code => {
-    if (isRestarting) return
-    process.exit(code)
-  })
+  developProcess.onExit(
+    (code: number | null, signal: NodeJS.Signals | null) => {
+      if (isRestarting) return
+      if (signal !== null) {
+        process.kill(process.pid, signal)
+        return
+      }
+      if (code !== null) {
+        process.exit(code)
+      }
+
+      // This should not happen:
+      // https://nodejs.org/api/child_process.html#child_process_event_exit
+      // The 'exit' event is emitted after the child process ends. If the process
+      // exited, code is the final exit code of the process, otherwise null.
+      // If the process terminated due to receipt of a signal, signal is the
+      // string name of the signal, otherwise null. One of the two will always be
+      // non - null.
+      //
+      // but just in case let do non-zero exit, because we are in situation
+      // we don't expect to be possible
+      process.exit(1)
+    }
+  )
 
   const rootFile = (file: string): string => path.join(program.directory, file)
 
   const files = [rootFile(`gatsby-config.js`), rootFile(`gatsby-node.js`)]
   let lastConfig = requireUncached(rootFile(`gatsby-config.js`))
-
-  let watcher
+  let watcher: chokidar.FSWatcher = null
 
   if (!isCI()) {
-    chokidar.watch(files).on(`change`, filePath => {
+    watcher = chokidar.watch(files).on(`change`, filePath => {
       const file = path.basename(filePath)
 
       if (file === `gatsby-config.js`) {
@@ -257,26 +402,68 @@ module.exports = async (program: IProgram): Promise<void> => {
     })
   }
 
-  process.on(`beforeExit`, async () => {
-    await Promise.all([
-      watcher?.close(),
-      unlock?.(),
-      new Promise(resolve => {
-        statusServer.close(resolve)
-      }),
-      new Promise(resolve => {
-        proxy.server.close(resolve)
-      }),
-    ])
+  // route ipc messaging to the original develop process
+  process.on(`message`, msg => {
+    developProcess.send(msg)
   })
 
   process.on(`SIGINT`, async () => {
-    await developProcess.stop(`SIGINT`)
-    process.exit()
+    await shutdownServices(
+      {
+        developProcess,
+        unlocks,
+        statusServer,
+        proxy,
+        watcher,
+      },
+      `SIGINT`
+    )
+
+    process.exit(0)
   })
 
   process.on(`SIGTERM`, async () => {
-    await developProcess.stop(`SIGTERM`)
-    process.exit()
+    await shutdownServices(
+      {
+        developProcess,
+        unlocks,
+        statusServer,
+        proxy,
+        watcher,
+      },
+      `SIGTERM`
+    )
+
+    process.exit(0)
   })
+
+  onExit((_code, signal) => {
+    shutdownServices(
+      {
+        developProcess,
+        unlocks,
+        statusServer,
+        proxy,
+        watcher,
+      },
+      signal as NodeJS.Signals
+    )
+  })
+}
+function shutdownServices(
+  { statusServer, developProcess, proxy, unlocks, watcher },
+  signal: NodeJS.Signals
+): Promise<void> {
+  const services = [
+    developProcess.stop(signal),
+    watcher?.close(),
+    new Promise(resolve => statusServer.close(resolve)),
+    new Promise(resolve => proxy.server.close(resolve)),
+  ]
+
+  unlocks.forEach(unlock => {
+    services.push(unlock())
+  })
+
+  return Promise.all(services).then(() => {})
 }