+++ packages/gatsby/src/services/customize-schema.ts
@@ -1,11 +1,12 @@
 import reporter from "gatsby-cli/lib/reporter"
 import { createSchemaCustomization } from "../utils/create-schema-customization"
-import { IBuildContext } from "./"
+import { IDataLayerContext } from "../state-machines/data-layer/types"
 
 export async function customizeSchema({
   parentSpan,
+  deferNodeMutation,
   refresh, // webhookBody,//coming soon
-}: Partial<IBuildContext>): Promise<void> {
+}: Partial<IDataLayerContext>): Promise<void> {
   const activity = reporter.activityTimer(`createSchemaCustomization`, {
     parentSpan,
   })
@@ -13,6 +14,7 @@ export async function customizeSchema({
   await createSchemaCustomization({
     parentSpan,
     refresh,
+    deferNodeMutation,
     // webhookBody,
   })
   activity.end()